Random Walk
We make a note of the start position (x,y), then we note whether the person walks right, left, forward, or back and add that to their position. We can then plot their position on a graph. We can just use Pythagoras’ theorem to work out the position (C file)