MM1 System
The simplest queueing theory model is the {M/M/1} model (which stands for Memoryless, Memoryless, and 1 server, and assumes: (a) that customers arrive in accordance with a Poisson process with average rate delta and thus the inter arrival times are exponentially distributed with mean 1/delta. Service time by the server is assumed exponential with parameter µ. Expected service time is then 1/µ. (c) The customers are served one at a time by a single server. If the server is busy upon the customer`s arrival, then the customer waits in the queue. (R file)